app/requirements.txt
chromadb==1.5.7
groq==1.1.2
pandas==3.0.2
python-dotenv==1.0.1
streamlit==1.56.0

